requirements.txt
# command line argument processing
cliff

# read openstack credentials from config file or environment
os_client_config

# openstack services
python-keystoneclient
python-novaclient
python-glanceclient
python-cinderclient

# We require python-swiftclient>3.1.0 for keystone session support
-e git+https://github.com/openstack/python-swiftclient/#egg=python-swiftclient
